+++ CHANGELOG.md
@@ -9,6 +9,16 @@ https://tapis.readthedocs.io/en/latest/technical/systems.html
 You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
+## 26Q1.1 - 2026-04-03
+
+Bug fix
+
+### New features:
+- None.
+
+### Bug fixes:
+- Fix bug with hostEval for the case of a dynamic effectiveUserId with a loginUser mapping.
+
 ## 26Q1.0 - 2026-03-04
 
 New features.